+++ src/go/plugin/go.d/collector/snmp/ddsnmp/ddsnmpcollector/collector_table.go
